fix(minimap): reduce log spam of inactive entities
**Problem:** We register all citizens to be rendered in the minimap when
they are spawned. For rendering an entity to the minimap we assume that
the location component is set on that entity and log a warning if it is
not.

However, the citizen entities are unloaded with the chunk if the player
moves away from a city. The citizen entities still stay registered, but
doing the lookup for the location component results in `null` for those
inactive entities. In consequence, we spam the log with WARN messages in
every `render` call:

```
00:06:50.569 [main] WARN  o.t.m.minimap.CharacterOverlay - Cannot render overlay for entity '103278' - missing LocationComponent
00:06:50.583 [main] WARN  o.t.m.minimap.CharacterOverlay - Cannot render overlay for entity '33707' - missing LocationComponent
00:06:50.583 [main] WARN  o.t.m.minimap.CharacterOverlay - Cannot render overlay for entity '103228' - missing LocationComponent
```

**Solution:** Check whether the entity is _active_ before attempting to
render it on the minimap. Also, add more information to the log message
in case we still run into that case. Additionally, add docstrings to a
few places to document the behavior and potential pitfalls.